refactor(#741): relocate geo_reverse.py to infrastructure/\n\nMove the Nominatim HTTP reverse-geocoder from the src/picframe top level to\ninfrastructure/, matching the documented Clean Architecture layer layout\n(geo_reverse is a network I/O adapter, not core domain logic).\n\n- git mv src/picframe/geo_reverse.py src/picframe/infrastructure/geo_reverse.py\n- Update import in core/services/geocoding_worker.py\n- No test changes needed (test_geocoding_worker mocks GeoReverse via fixture)\n\nGates: ruff, ruff format, mypy, pytest 801 passed.\n\nCloses #741
